move options out + location change
