+++ packages/commons-ui-core/styles/createTheme.js
@@ -1,0 +1,7 @@
+import { createTheme } from "@mui/material/styles";
+
+const customTheme = (options) => {
+  return createTheme(options);
+};
+
+export default customTheme;
